scripts/requirements.txt
mysql-connector-python>=8.0.33
bcrypt>=4.1.0
python-dotenv>=1.0.0
